+++ Assets/Scripts/SelfClean.cs
@@ -1,0 +1,10 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SelfClean : MonoBehaviour {
+    void Update() {
+        if (transform.childCount == 0)
+            Destroy(gameObject, 2f);
+    }
+}