+++ Assets/Scripts/TrackerShortMovement.cs
@@ -2,24 +2,21 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-/*For Red Tracking Enemy: Move for a short time, then stop (and start acting and attacking)*/
-
+/*For Red Tracking Enemy: Move for a short time, then Track Attack (while continue Moving)*/
 public class TrackerShortMovement : MonoBehaviour
 {
     public float startDelay;
     public float moveSpeed;
-
     public Vector2 timeRange;
 
     Rigidbody2D rgbd;
 
     void Start() {
         rgbd = GetComponent<Rigidbody2D>();
-
         StartCoroutine("MoveShort");
     }
 
-    //Move for a set amount of time then stop:
+    //Move for a set amount of time then start tracking:
     IEnumerator MoveShort() {
         yield return new WaitForSeconds(startDelay);
 
@@ -29,9 +26,7 @@ public class TrackerShortMovement : MonoBehaviour
 
         yield return new WaitForSeconds(moveTime);
 
-        rgbd.velocity = new Vector2(0, 0);
-
-        //After finish moving, start rotating:
+        // Strat Tracking after sometime moving
         StartTracking();
 
         yield break;